Move admin toggler to manage employee view
